+++ mainnet_launch/pages/rebalance_events/rebalance_events.py
@@ -34,7 +34,7 @@ from mainnet_launch.data_fetching.add_info_to_dataframes import (
     add_timestamp_to_df_with_block_column,
     add_transaction_gas_info_to_df_with_tx_hash,
 )
-from mainnet_launch.database.new_databases import (
+from mainnet_launch.database.database_operations import (
     write_dataframe_to_table,
     run_read_only_query,
     get_earliest_block_from_table_with_autopool,
