+++ pyproject.toml
@@ -19,7 +19,6 @@ python-dotenv = "^1.0.1"
 streamlit = "1.38.0"
 boto3 = "^1.35.29"
 psutil = "^6.0.0"
-portalocker = "^3.1.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.3"
